test_requirements.txt
pytest==3.2.2
pytest-django==3.1.2
pytest-cov==2.5.1
